Rename function files.
